rename components ant titles 'Day' to 'Current'
