+++ src/components/Menu/Menu.jsx
@@ -23,12 +23,12 @@ const MenuFC = () => {
       >
         <Menu.Item key='today' icon={<CloudOutlined />}>
           <NavLink className={({ isActive }) => (isActive ? s.nav__link_active : s.nav__link)} to='/current-weather'>
-            Today
+            Current
           </NavLink>
         </Menu.Item>
         <Menu.Item key='week' icon={<CompassOutlined />}>
           <NavLink className={({ isActive }) => (isActive ? s.nav__link_active : s.nav__link)} to='/weather-forecast'>
-            Week
+            Forecast
           </NavLink>
         </Menu.Item>
         <Menu.Item key='dashboard' icon={<DashboardOutlined />}>
