+++ src/pages/Week/Week.module.css
@@ -1,8 +1,4 @@
 .wrapper {
-  /* display: flex; */
-  /* justify-content: center; */
   margin-bottom: 2em;
-  background-color: green;
-  /* text-align: center; */
-  margin: 0 auto;
+  text-align: center;
 }
